Imported Trash Dash + Changed Lives to Five (5)
Changed the local variable in the "GameState" script to  the value expressed in the character controller. Changed the value in the character controller to 5. 
Reworked the lives ui to account for the new maximum.
Changed the numeric value in the "Hearts" powerup to correctly work with a maximum of 5 lives.